apps/bot-whatsapp/requirements.txt
fastapi
uvicorn
requests
boto3
python-dotenv
python-dateutil
google-generativeai
pydantic
